Rename relative_attitude_controller
