+++ s2e-ff/src/components/ideal/initialize_relative_attitude_controller.cpp
@@ -1,4 +1,4 @@
-#include "InitializeRelativeAttitudeController.hpp"
+#include "initialize_relative_attitude_controller.hpp"
 
 #include <library/initialize/initialize_file_access.hpp>
