+++ s2e-ff/src/components/ideal/relative_attitude_controller.hpp
