+++ s2e-ff/src/simulation/spacecraft/ff_components.hpp
@@ -11,7 +11,7 @@
 #include "../../components/aocs/initialize_relative_distance_sensor.hpp"
 #include "../../components/aocs/initialize_relative_position_sensor.hpp"
 #include "../../components/aocs/initialize_relative_velocity_sensor.hpp"
-#include "../../components/ideal/InitializeRelativeAttitudeController.hpp"
+#include "../../components/ideal/initialize_relative_attitude_controller.hpp"
 
 class FfComponents : public InstalledComponents {
  public:
